# docs/swagger.yaml
openapi: 3.0.0

info:
  title: docs
  description: 'learn swagger and express'
  version: 1.0.0
  contact:
    email: abhishek.sankpal12@gmail.com
    url: 'https://facke.com'

servers:
  - url: '{protocol}://localhost:4000/api/{version}'
    description: Expermintal Production Server
    variables:
      version:
        default: v1
        enum:
          - v1
          - v2
      protocol:
        default: http
        enum:
          - http
          - https

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer

paths:
  /greet:
    get:
      tags:
        - string
      summary: Return Welcome Message.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: string
                example: 'mystring'

        '400':
          description: Bad Request
        '500':
          description: internal server error

  /songsobj:
    get:
      tags:
        - Object
      summary: Return list of songs.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: string

        '400':
          description: Bad Request
        '500':
          description: internal server error

  /songarr:
    get:
      tags:
        - Arrays
      summary: Return list of all songs.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              items:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  price:
                    type: string

        '400':
          description: Bad Request
        '500':
          description: internal server error

  /songs/{id}:
    get:
      tags:
        - Object
      summary: Get a userby Id
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  rating:
                    type: string

        '400':
          description: Bad Request
        '500':
          description: internal server error

  #  Create Route
  /createSong:
    post:
      tags:
        - Object
      summary: Create Song
      consumes:
        - application/json
      produces:
        - application/json
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                rating:
                  type: string

      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  name:
                    type: string
                  rating:
                    type: string

        '400':
          description: Bad Request
        '500':
          description: internal server error

  #  Search Route
  /searchSong:
    get:
      tags:
        - Object
      summary: search song
      parameters:
        - name: song
          in: query
          required: true
          schema:
            type: string
            enums: [pune, banglore, ohio]

        - name: rating
          in: query
          required: true
          schema:
            type: string
            enums: [10, 5.8, 7]
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    type: string
                  rating:
                    type: string

        '400':
          description: Bad Request
        '500':
          description: internal server error